chore: cleanup before rewrite
