+++ src-tauri/Cargo.toml
@@ -11,15 +11,15 @@ rust-version = "1.57"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [build-dependencies]
-tauri-build = { version = "1.2", features = [] }
+tauri-build = { version = "1.4.0", features = [] }
 
 [dependencies]
 serde = "1.0.152"
 serde_derive = "1.0.152"
 serde_json = "1.0.91"
-tauri = { version = "1.2.4", features = ["api-all", "system-tray"] }
+tauri = { version = "1.4.1", features = ["api-all", "system-tray"] }
 idasen = "0.2.0"
-btleplug = "0.10.4"
+btleplug = "0.11.0"
 tokio = { version = "1.24.1", features = ["full"] }
 indicatif = "0.16.2"
 thiserror = "1.0.30"
