+++ src-tauri/src/loose_idasen.rs
@@ -77,6 +77,24 @@ where
     pub position_characteristic: Characteristic,
 }
 
+pub async fn get_list_of_desks(loc_name: &Option<String>) -> Vec<ExpandedPeripheral> {
+    let desks = match loc_name {
+        // If local name was provided
+        Some(loc_name) => {
+            let desks = get_desks(Some(loc_name.clone())).await;
+            desks
+        }
+        // If local name was NOT provided
+        None => {
+            let desks = get_desks(None).await;
+            desks
+        }
+    };
+    let desks = desks.expect("Error while getting a list of desks");
+
+    desks
+}
+
 /// Do a set of tasks for a peripheral to make it usable.
 pub async fn setup(desk: &impl ApiPeripheral) -> Result<Idasen<impl ApiPeripheral>, Error> {
     let mac_addr = desk.address();
@@ -109,8 +127,8 @@ pub async fn setup(desk: &impl ApiPeripheral) -> Result<Idasen<impl ApiPeriphera
     })
 }
 
-/// Ggetting characteristics every time is wasteful
-/// TODO: Try to refactor this - maybe chuck this into shared tauri state?
+// Getting characteristics every time is wasteful
+// TODO: Try to refactor this - maybe chuck this into shared tauri state?
 pub async fn get_control_characteristic(desk: &impl ApiPeripheral) -> Characteristic {
     desk.characteristics()
         .iter()
@@ -119,8 +137,7 @@ pub async fn get_control_characteristic(desk: &impl ApiPeripheral) -> Characteri
         .expect("err while getting characteristic")
         .clone()
 }
-/// Getting characteristics every time is wasteful
-/// TODO: Try to refactor this - maybe chuck this into shared tauri state?
+
 pub async fn get_position_characteristic(desk: &impl ApiPeripheral) -> Characteristic {
     desk.characteristics()
         .iter()
@@ -130,31 +147,26 @@ pub async fn get_position_characteristic(desk: &impl ApiPeripheral) -> Character
         .clone()
 }
 
-pub async fn up(desk: &impl ApiPeripheral) -> btleplug::Result<()> {
+async fn up(desk: &impl ApiPeripheral) -> btleplug::Result<()> {
     let control_characteristic = get_control_characteristic(desk).await;
 
     desk.write(&control_characteristic, &UP, WriteType::WithoutResponse)
         .await
 }
 
-pub async fn down(desk: &impl ApiPeripheral) -> btleplug::Result<()> {
+async fn down(desk: &impl ApiPeripheral) -> btleplug::Result<()> {
     let control_characteristic = get_control_characteristic(desk).await;
     desk.write(&control_characteristic, &DOWN, WriteType::WithoutResponse)
         .await
 }
 
-pub async fn stop(desk: &impl ApiPeripheral) -> btleplug::Result<()> {
+async fn stop(desk: &impl ApiPeripheral) -> btleplug::Result<()> {
     let control_characteristic = get_control_characteristic(desk).await;
     desk.write(&control_characteristic, &STOP, WriteType::WithoutResponse)
         .await
 }
 
-/// Move desk to a position
-pub async fn move_to(desk: &impl ApiPeripheral, target_position: u16) -> Result<(), Error> {
-    move_to_target(desk, target_position).await
-}
-
-async fn move_to_target(desk: &impl ApiPeripheral, target_position: u16) -> Result<(), Error> {
+pub async fn move_to_target(desk: &impl ApiPeripheral, target_position: u16) -> Result<(), Error> {
     println!("starting moving to target");
     if !(MIN_HEIGHT..=MAX_HEIGHT).contains(&target_position) {
         return Err(Error::PositionNotInRange);
